src/main/java/com/example/leetCode/_121.java
package com.example.leetCode;//You are given an array prices where prices[i] is the price of a given stock on
// the ith day.
//
// You want to maximize your profit by choosing a single day to buy one stock an
//d choosing a different day in the future to sell that stock.
//
// Return the maximum profit you can achieve from this transaction. If you canno
//t achieve any profit, return 0.
//
//
// Example 1:
//
//
//Input: prices = [7,1,5,3,6,4]
//Output: 5
//Explanation: Buy on day 2 (price = 1) and sell on day 5 (price = 6), profit =
//6-1 = 5.
//Note that buying on day 2 and selling on day 1 is not allowed because you must
// buy before you sell.
//
//
// Example 2:
//
//
//Input: prices = [7,6,4,3,1]
//Output: 0
//Explanation: In this case, no transactions are done and the max profit = 0.
//
//
//
// Constraints:
//
//
// 1 <= prices.length <= 105
// 0 <= prices[i] <= 104
//
// Related Topics Array Dynamic Programming
// 👍 7453 👎 330


//leetcode submit region begin(Prohibit modification and deletion)
class _121 {
  public int maxProfit(int[] prices) {
    int min = prices[0];
    int max = 0;
    int abs = 0;
    for (int price : prices) {
      if (price <= min) {
        abs = Math.max(Math.max(0, max - min), abs);

        min = price;
        max = price;
        continue;
      }

      if (price >= max) {
        max = price;
        abs = max - min;
      }
    }

    return Math.max(abs, max - min);
  }
}
//leetcode submit region end(Prohibit modification and deletion)
